POD-92: Statistics page

See merge request podbike-developers/podbike-app-android!1

+++ app/src/main/java/com/podbike/app/ui/dashboard/DeviceDataUiModel.kt
@@ -13,7 +13,7 @@ data class DeviceDataUiModel(
     val lightStatus: PodbikeLightStatus,
     val distanceAbbreviation: String,
     val range: String,
-    val temperature: Int,
+    val temperature: Float,
     val isMoving: Boolean,
     val name: String?
 )
